Fix export leaking unrounded PSNR; centralize sig-fig formatting (#4422)
## Summary

`qai-hub-models export` was printing PSNR at full float precision while
every other metric in the same table was correctly truncated to 4
significant figures:

```
psnr: 39.953045   <- 8 sig figs, unformatted
mse:  0.0001016   <- correctly .4g
mae:  0.007999    <- correctly .4g
```

**Root cause:** `compute_psnr` returned `20 * np.log10(...)` computed on
float32 inputs — an `np.float32`. numpy scalars are not `float`
subclasses (except `float64`), so the `isinstance(x, float)` check in
`print_inference_metrics.custom_float_format` silently skipped the psnr
column and passed the raw value through.

Every other metric function in `compare.py` already wraps its return in
`float()`. `compute_psnr` was the lone omission, despite its `-> float`
annotation.

## Changes

**`utils/compare.py`** — wrap `compute_psnr`'s return in `float()`. This
is the actual fix; it corrects the export path at the source.

**`utils/printing.py`** — the sig-fig policy was duplicated as bare
`.4g`/`.3g` literals across four call sites, which is why the scorecard
path looked clean (it had been patched separately in #3081) while export
did not. Those literals now come from one place:

```python
METRIC_SIG_FIGS = 4       format_metric()     # PSNR, MSE, MAE, ...
ACCURACY_SIG_FIGS = 3     format_accuracy()   # accuracy percentages
format_sig_figs(value, sig_figs)
is_float(value)                               # TypeGuard[float], also accepts np.floating
```

`custom_float_format` now tests via `is_float`, so a future numpy-scalar
leak cannot bypass rounding the way this one did.

**Three call sites** now import these instead of carrying their own
literals:
- `scorecard/utils/testing.py` — `mock_tabulate_fn`
- `scorecard/static/scripts/evaluate_static_accuracy.py`
- `scorecard/utils/testing_async_utils.py` — `write_accuracy`'s three
accuracy fields

No new module: the helpers live next to the code that uses them. There's
no import cycle — `compare.py` needs only the bare `float()` cast, not
the formatters.

## Not included

- **`accuracy.csv` is unchanged.** Auditing the committed file, **0 of
5,985** values ever exceeded their sig-fig budget — the truncation was
never lost on the scorecard side, only on export.
- A pandas `read_csv`/`to_csv` round-trip in
`collect_scorecard_results.py` re-renders integer-valued entries as
`100.0` (398 values). Confirmed cosmetic — no precision loss — and left
as-is per review.
- `_display_with_sig_figs` in `autofill_manifest_yaml.py` is untouched;
it's fixed-decimal file-size formatting (`1.50 MB`), a different
concern.

## Test plan

- [x] `compute_psnr` returns a real `float`; export table renders PSNR
as `40.17`
- [x] `is_float` accepts `np.float32`/`np.float64`/`float`, rejects
tuples and strings
- [x] All six affected modules import cleanly (no cycle from
`testing_async_utils` → `printing`)
- [x] 585 passed, 149 skipped — `scorecard/test/`,
`test_numerics_yamls.py`, `test_utils/`
- [x] `pre-commit` clean on all touched files (ruff, ruff-format,
pydoclint, mypy)

+++ src/qai_hub_models/utils/compare.py
@@ -119,7 +119,9 @@ def compute_psnr(
 
     data_range_est = data_range if data_range is not None else np.abs(b).max()
 
-    return 20 * np.log10((data_range_est + eps) / (sumdeltasq + eps2))
+    # float() so callers get a real Python float, matching every other metric
+    # here; a leaked np.float32 bypasses isinstance(x, float) formatting.
+    return float(20 * np.log10((data_range_est + eps) / (sumdeltasq + eps2)))
 
 
 def compute_relative_error(expected: np.ndarray, actual: np.ndarray) -> np.ndarray:
